feat: filters save in localstorage
