+++ .eslintrc
@@ -13,6 +13,7 @@
   "parserOptions": { "ecmaVersion": "latest", "sourceType": "module" },
   "settings": { "react": { "version": "18.2" } },
   "plugins": ["react-refresh", "unused-imports", "vitest"],
+  "parser": "@babel/eslint-parser",
   "rules": {
     "react-refresh/only-export-components": [
       "warn",
