+++ .nvmrc
@@ -1,1 +1,1 @@
-v18.18.0
+v20.6.0
