+++ src/__tests__/hooks/useUrlValues.test.jsx
@@ -5,7 +5,7 @@ import {
   DEFAULT_PAGE,
   DEFAULT_SORTING_TAG,
 } from '@/lib/utils/config';
-import UrlProvider, { useUrlValues } from '@/providers/urlProvider';
+import UrlProvider, { useUrlValues } from '@/providers/urlProvider/reducer';
 import { act, render } from '@testing-library/react';
 import { beforeEach, describe, expect, it } from 'vitest';
